+++ src/cloregram/impl/resources.clj
@@ -1,0 +1,18 @@
+(ns cloregram.impl.resources
+  (:require [clojure.java.io :as io]
+            [clojure.edn :as edn]
+            [clojure.string :as str]
+            [resauce.core :as res]))
+
+(defn- read-resource [resource-url]
+  (with-open [stream (io/input-stream resource-url)]
+    (-> stream
+        io/reader
+        java.io.PushbackReader. edn/read)))
+
+(defn read-resource-dir
+  [dir]
+  (when-let [resources (some-> dir io/resource res/url-dir)]
+    (->> resources
+         (filter #(str/ends-with? % ".edn"))
+         (mapcat read-resource))))